fileserver/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
openpyxl
numpy
pandas
gunicorn
